Added check fo ifr author is teaching a course before allowing them to be deleted.

+++ building-react-flux-apps/psadmin/src/actions/authorActions.js
@@ -3,6 +3,8 @@
 var Dispatcher = require('../dispatcher/appDispatcher');
 var AuthorApi = require('../api/authorApi');
 var ActionTypes = require('../constants/actionTypes');
+var CourseStore = require('../stores/courseStore');
+var toastr = require('toastr');
 
 var AuthorActions = {
     createAuthor: function(author) {
@@ -30,14 +32,34 @@ var AuthorActions = {
     },
 
     deleteAuthor: function(id) {
-        AuthorApi.deleteAuthor(id);
+        var coursesWithAuthorId = CourseStore.getCoursesByAuthorId(id);
 
+        if (coursesWithAuthorId.length > 0) {
+            // Don't let the user delete an author who is listed as teaching one or more courses.
+            toastr.error(
+                "Can not delete because this author is teaching the following courses:<br/>\n" +
+                "<ol>\n" +
+                coursesWithAuthorId.map(
+                    function(course) {
+                        return ("<li>" + course.title + "</li>\n");
+                    }
+                ).join('') +
+                "\n" +
+                "</ol>\n"
+            );
+
+            return false;
+        }
+
+        AuthorApi.deleteAuthor(id);
         Dispatcher.dispatch(
             {
                 actionType: ActionTypes.DELETE_AUTHOR,
                 id: id
             }
         );
+
+        return true;
     }
 };
